+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. fconvert documentation master file, created by
+   sphinx-quickstart on Sun Mar 12 16:39:29 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to fconvert's documentation!
+====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
