+++ src/fconvert.egg-info/SOURCES.txt
@@ -3,6 +3,7 @@ README.md
 pyproject.toml
 src/fconv/__init__.py
 src/fconv/fconvert.py
+src/fconv/utils.py
 src/fconvert.egg-info/PKG-INFO
 src/fconvert.egg-info/SOURCES.txt
 src/fconvert.egg-info/dependency_links.txt
